Refactor SecretItem class and update usage in Program.cs
Updated the `SecretItem` class to include XML documentation and replaced the `FileName` property with `ProjectFileName` to store the full path of the associated project file. Enhanced documentation for the `UserSecretsId` property. Modified the instantiation of `SecretItem` in `Program.cs` to use the new property and removed obsolete logging code.

by Karen